Spliting instruction representation in 2
An instruction can be represented as Encoded (on 64 bits) or Decoded in a struct

+++ src/sponge_std/sponge_testing.h
@@ -32,6 +32,6 @@ void pass();
 
 void printResult(char *testName, Test testResult);
 
-void printAsBinary(uint32_t integer);
+void printAsBinary(uint64_t integer);
 
 #endif
